fix: make default env and org initializers upgraders
