+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/upgrade/initializer/EnvironmentInitializer.java
@@ -37,7 +37,7 @@ public abstract class EnvironmentInitializer implements Initializer {
     @Autowired
     private EnvironmentRepository environmentRepository;
 
-    // upgradeEnvironment is called once for each environment
+    // initializeEnvironment is called once for each environment
     protected abstract void initializeEnvironment(ExecutionContext executionContext);
 
     @Override
