+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/test/java/io/gravitee/rest/api/service/impl/upgrade/upgrader/DefaultEnvironmentUpgraderTest.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.service.impl.upgrade.initializer;
+package io.gravitee.rest.api.service.impl.upgrade.upgrader;
 
 import static org.mockito.Mockito.*;
 
-import io.gravitee.rest.api.service.OrganizationService;
-import io.gravitee.rest.api.service.impl.upgrade.initializer.DefaultOrganizationInitializer;
+import io.gravitee.rest.api.service.EnvironmentService;
+import java.util.List;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
@@ -29,25 +29,20 @@ import org.mockito.junit.MockitoJUnitRunner;
  * @author GraviteeSource Team
  */
 @RunWith(MockitoJUnitRunner.class)
-public class DefaultOrganizationInitializerTest {
+public class DefaultEnvironmentUpgraderTest {
 
     @Mock
-    private OrganizationService organizationService;
+    private EnvironmentService environmentService;
 
     @InjectMocks
-    private final DefaultOrganizationInitializer initializer = new DefaultOrganizationInitializer();
+    private final DefaultEnvironmentUpgrader upgrader = new DefaultEnvironmentUpgrader();
 
     @Test
-    public void shouldInitializeOrganisation() {
-        when(organizationService.count()).thenReturn(0L);
-        initializer.initialize();
-        verify(organizationService, times(1)).initialize();
-    }
+    public void shouldCreateDefaultEnvironment() {
+        when(environmentService.findByOrganization(any())).thenReturn(List.of());
 
-    @Test
-    public void shouldNotInitializeOrganisation() {
-        when(organizationService.count()).thenReturn(1L);
-        initializer.initialize();
-        verify(organizationService, never()).initialize();
+        upgrader.upgrade();
+
+        verify(environmentService, times(1)).initialize();
     }
 }
